pyproject.toml
[project]
name = "fraud-detection-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-beam>=2.70.0",
    "avro-python3>=1.10.2",
    "faker>=40.15.0",
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage>=3.10.1",
    "kafka-python-ng>=2.2.3",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "rich>=15.0.0",
    "scikit-learn>=1.8.0",
    "ulid-py>=1.1.0",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
